Fix #3 Add verification to config
Configurations are now loaded into a temporary variable, prior to being
written into the actual configuration location. Between those two steps,
the configuration is verified for any illegal settings, and failing at
it prevents writing the actual variable.

The verification syntax seems overly complex at the moment, but it
allows us a simple implementation of verification by each subsystem when
the need arises. With this syntax, failure of verification for one
system doesn't stop the verification process, but continues it, so we
can detect all issues, and not only the first one detected.