Improve logging speed for the SD card
Saving full state messages to the buffer takes ~60us, and sending 512B
chunks to the SD card takes additional ~200us, resulting in state
messages needing ~253us in ~45% of the cases, and ~453us in the rest of
the cases. That includes ~173us needed to process the packages and send
them via serial.